+++ tests/Extractor/Fixtures/Foo.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AutoMapper\Tests\Extractor\Fixtures;
+
+class Foo
+{
+    public string $bar;
+    public string $baz;
+}